# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office Planner Suite API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  status?: number;
}

export interface CatalogItem {
  id: string;
  name: string;
  category: string;
  widthCm: number;
  depthCm: number;
  heightCm: number;
  /** @nullable */
  color: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  shape: string | null;
  /** @nullable */
  seatCount: number | null;
  /** @nullable */
  price: number | null;
}

export interface CategorySummary {
  name: string;
  count: number;
}

export type PlannerType = (typeof PlannerType)[keyof typeof PlannerType];

export const PlannerType = {
  canvas: "canvas",
  blueprint: "blueprint",
  floorplan: "floorplan",
  cad: "cad",
  shapes: "shapes",
  import: "import",
  "oando-floor-planning": "oando-floor-planning",
  "oando-cad-drawing": "oando-cad-drawing",
  "oando-site-plan": "oando-site-plan",
  "oando-floor-plan-creator": "oando-floor-plan-creator",
  "oando-custom-shapes": "oando-custom-shapes",
  "oando-import-scale": "oando-import-scale",
  "oando-templates": "oando-templates",
} as const;

export interface PlanSummary {
  id: number;
  name: string;
  plannerType: PlannerType;
  roomWidthCm: number;
  roomDepthCm: number;
  itemCount: number;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Plan {
  id: number;
  name: string;
  plannerType: PlannerType;
  roomWidthCm: number;
  roomDepthCm: number;
  documentJson: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePlanBody {
  name: string;
  plannerType: PlannerType;
  roomWidthCm: number;
  roomDepthCm: number;
  documentJson: string;
  /** @nullable */
  projectId?: number | null;
}

export interface UpdatePlanBody {
  name?: string;
  roomWidthCm?: number;
  roomDepthCm?: number;
  documentJson?: string;
  /** @nullable */
  projectId?: number | null;
}

export interface DuplicatePlanBody {
  name?: string;
}

export interface PlanStats {
  totalPlans: number;
  canvas2dPlans: number;
  blueprintPlans: number;
  /** @nullable */
  recentPlan: string | null;
}

export interface SpatialItem {
  x: number;
  y: number;
  widthCm: number;
  depthCm: number;
  rotation: number;
  name: string;
}

export interface AiAdvisorRequest {
  roomWidthCm: number;
  roomDepthCm: number;
  itemCount?: number;
  query: string;
  categories?: string[];
  items?: SpatialItem[];
}

export interface AiAdvisorResponse {
  advice: string;
  suggestions: string[];
  issues?: string[];
  positives?: string[];
}

export interface TemplateSummary {
  id: string;
  name: string;
  description: string;
  category: string;
  roomWidthCm: number;
  roomDepthCm: number;
  layoutJson: string;
  furnitureCount: number;
  usageCount: number;
  /** @nullable */
  thumbnailSvg?: string | null;
  createdAt: string;
}

export interface UseTemplateBody {
  name?: string;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientWithStats {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  projectCount: number;
  createdAt: string;
}

export interface CreateClientBody {
  name: string;
  company?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface UpdateClientBody {
  name?: string;
  company?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  clientId?: number | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectWithStats {
  id: number;
  name: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  status: string;
  /** @nullable */
  notes?: string | null;
  planCount: number;
  createdAt: string;
  updatedAt: string;
}

/**
 * @nullable
 */
export type ProjectDetailClient = {
  id?: number;
  name?: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
} | null;

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  client?: ProjectDetailClient;
  status: string;
  /** @nullable */
  notes?: string | null;
  plans: PlanSummary[];
  planCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  clientId?: number;
  status?: string;
  notes?: string;
}

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  clientId?: number | null;
  status?: string;
  notes?: string;
}

export type ListCatalogItemsParams = {
  category?: string;
  search?: string;
};

export type ListPlansParams = {
  limit?: number;
  offset?: number;
};

export type ListTemplatesParams = {
  category?: string;
};

export type ListClientsParams = {
  search?: string;
};

export type ListProjectsParams = {
  clientId?: number;
  status?: string;
};