# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Office Planner Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      security:
        - BearerAuth: []
      summary: List all saved floor plans
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of floor plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanSummary"
    post:
      operationId: createPlan
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Create a new floor plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanBody"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Get a floor plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A floor plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePlan
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Update a floor plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanBody"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlan
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Delete a floor plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/{id}/duplicate:
    post:
      operationId: duplicatePlan
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Duplicate an existing floor plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DuplicatePlanBody"
      responses:
        "201":
          description: Duplicated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Source plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/stats:
    get:
      operationId: getPlanStats
      tags: [plans]
      security:
        - BearerAuth: []
      summary: Get plan statistics summary
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanStats"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      security:
        - BearerAuth: []
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientWithStats"
    post:
      operationId: createClient
      tags: [clients]
      security:
        - BearerAuth: []
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      security:
        - BearerAuth: []
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithStats"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      security:
        - BearerAuth: []
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      security:
        - BearerAuth: []
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
      required:
        - error
...
    PlannerType:
      type: string
      enum:
        - canvas
        - blueprint
        - floorplan
        - cad
        - shapes
        - import
        - oando-floor-planning
        - oando-cad-drawing
        - oando-site-plan
        - oando-floor-plan-creator
        - oando-custom-shapes
        - oando-import-scale
        - oando-templates
    PlanSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        plannerType:
          $ref: "#/components/schemas/PlannerType"
        roomWidthCm:
          type: number
        roomDepthCm:
          type: number
        itemCount:
          type: integer
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - plannerType
        - roomWidthCm
        - roomDepthCm
        - itemCount
        - createdAt
        - updatedAt
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        plannerType:
          $ref: "#/components/schemas/PlannerType"
        roomWidthCm:
          type: number
        roomDepthCm:
          type: number
        documentJson:
          type: string
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - plannerType
        - roomWidthCm
        - roomDepthCm
        - documentJson
        - createdAt
        - updatedAt
    CreatePlanBody:
      type: object
      properties:
        name:
          type: string
        plannerType:
          $ref: "#/components/schemas/PlannerType"
        roomWidthCm:
          type: number
        roomDepthCm:
          type: number
        documentJson:
          type: string
        projectId:
          type: ["integer", "null"]
      required:
        - name
        - plannerType
        - roomWidthCm
        - roomDepthCm
        - documentJson
    UpdatePlanBody:
      type: object
      properties:
        name:
          type: string
        roomWidthCm:
          type: number
        roomDepthCm:
          type: number
        documentJson:
          type: string
        projectId:
          type: ["integer", "null"]
    DuplicatePlanBody:
      type: object
      properties:
        name:
          type: string
    PlanStats:
      type: object
      properties:
        totalPlans:
          type: integer
        canvas2dPlans:
          type: integer
        blueprintPlans:
          type: integer
        recentPlan:
          type: ["string", "null"]
      required:
        - totalPlans
        - canvas2dPlans
        - blueprintPlans
        - recentPlan
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    ClientWithStats:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        projectCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - projectCount
        - createdAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
      required:
        - name
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
...