# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office Planner Suite API
 * OpenAPI spec version: 0.2.0
 */

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}