# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Office Planner Suite API
 * OpenAPI spec version: 0.2.0
 */
import type { PlannerType } from "./plannerType";

export interface Plan {
  id: number;
  name: string;
  plannerType: PlannerType;
  roomWidthCm: number;
  roomDepthCm: number;
  documentJson: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
  updatedAt: string;
}